Adding additional overloadings for CKInsetComponent
Summary:
As a follow up for D25613073, we want to add an additional overloadings to cover both `CKRelativeDimension` and `CGFloat` for all the builders.

The reason is:
- When you have a more specialised overload the conversion to `CKRelativeDimension` happens inside the builder method, not at the call site.
- When you have a single setter, the conversion is emitted at the call site.

If the setter is not inlined having two different overloads results in binary size savings (as the conversion code is not duplicated at each call site).

Reviewed By: cuva

Differential Revision: D25900713

fbshipit-source-id: d7b88e684f6033682f39050110f5b87365691892