Added FitContent size provider
Summary:
None of the other options in `CKComponentSizeRangeFlexibility` allow the root component to determine the size of the hosting view:

- `CKComponentSizeRangeFlexible*` are only suitable for scrolling containers since they don't constrain the size in the scrolling direction.
- `CKComponentSizeRangeFlexibilityNone` always stretches the root component to the size of the hosting view.

The new option, `CKComponentSizeRangeFitContent`, allows the root component to be any size it wants while still being bounded by the size of the hosting view itself (which is good since `CKComponentHostingView` cannot scroll). To me, this sounds like an option that makes the most sense for the vast majority of hosting views.

Reviewed By: kfirapps, cuva

Differential Revision: D26691120

fbshipit-source-id: a7e0dd05ce4185b01dd9b5f781306bd4a65041ab