Remove conditional scope handle creation in `CKTreeNode`
Summary:
`CKTreeNode` used to check if `component.scopeHandle != nil` before creating its own scope handle. This doesn't seem to be possible anymore since `-scopeHandle` is a getter on `CKComponent` which returns `treeNode.scopeHandle` unless the component we're creating a tree node for already has a tree node - in which case we have bigger problems.

Used to be part of next diff but decided to extract it to make this change clearer.

Reviewed By: kfirapps, alickbass

Differential Revision: D26635666

fbshipit-source-id: 976f2d0de03c649fe2e21258b211f2fe2241c6a1