fixed cache leak when context contains non-recreatable/invalid context
Summary: This change allows reuse pool to "ask" controller if context is still valid and can be recreated for future reuse. Goal is to fix cache leak issue, when view is stored in reuse pool but would be never reused, since context contains piece of data that can't be recreated anymore.

Reviewed By: kevin0571

Differential Revision: D26914733

fbshipit-source-id: de4247b78328631b9407a9f5a844a3d535e4972c